Add files via upload
Inheritance is a fundamental concept in object-oriented programming (OOP) that allows a class (subclass or derived class) to inherit properties and behaviors (methods and fields) from another class (superclass or base class). Inheritance facilitates code reuse, extensibility, and the creation of a hierarchical structure in the code.

Here are some key concepts related to inheritance:

Superclass/Base Class:

The class whose properties and behaviors are inherited by another class is called the superclass or base class. It serves as the template or blueprint for creating new classes.
Subclass/Derived Class:

The class that inherits properties and behaviors from a superclass is called the subclass or derived class. The subclass can also have additional or overridden methods and fields.