Add files via upload
Association is a relationship between two or more classes, where objects of one class are related to objects of another class. This relationship can be one-to-one, one-to-many, or many-to-many. Associations are established through the use of class attributes or methods that reference other classes.
For example, consider a Teacher class and a Student class. There is an association between them, as teachers teach students. This relationship might be represented by a method like Teach in the Teacher class that takes a Student object as a parameter.


capsulation is one of the fundamental principles of object-oriented programming that involves bundling the data (attributes) and the methods (functions) that operate on the data into a single unit called a class. It also involves restricting access to some of the object's components.
Encapsulation helps in hiding the internal details of an object and exposing only what is necessary. It provides data abstraction, making it easier to manage the complexity of a system and promoting modularity.

Function Overloading refers to the ability to define multiple functions in the same scope with the same name but different parameters. The function to be called is determined at compile-time based on the number or types of parameters.